refactor(metrics): extract npa/loc shared helpers below 800-SLOC gate
After the per-language split (#969 precedent), npa.rs and loc.rs still
sat over the project's 800-SLOC self-scan limit (950 and 827), held
green only by `.bca-baseline.toml`. The residue was the irreducible
shared core: the public `Stats` struct plus a cluster of cross-language
`pub(crate)` helper functions shared by the per-language submodules and
sibling metrics.

Move those helpers into `npa/shared.rs` and `loc/shared.rs` (each
`use super::*;`) and re-export them with `pub(crate) use shared::*;` so
every existing call site resolves unchanged. The public `Stats`, the
`Npa` / `Loc` traits, `min_or_zero`, and the per-language dispatch
macros stay in the parent. `rust_visibility_modifier_is_public` (a
former parent-private fn reached by a sibling module) widens to
`pub(crate)` so the re-export carries it.

Pure move: no public-API change (src/lib.rs byte-identical) and no
metric-output change (zero integration-snapshot churn). npa.rs drops to
550 SLOC, loc.rs to 719; both file-level baseline entries retire.

Fixes #976
